Compare symbols instead of names in DeCapturifyTransformer
The DeCapturifyTransformer would fail to match `This` tree node
because it tried to compare names of qualifiers. However, names
can be mangled by flatten phase. This tranformation seems to be
not reflected in name stored in `This` node. Which is a bug in
flatten phase.

Comparing Symbols is usually more reliable.