Pass symbols for parameters to MethodType.
Fix construction of the MethodType corresponding to a constructor
in Delambdafy. Instead of passing a symbol corresponding to local
val def for a parameter there was a symbol corresponding to type
of that val def being passed. An example might explain the
difference. Let's assume we have a method:

def m(x: Array[Int]): Int

then it's MethodType should look like this:

MethodType(List(<symbol of x>), IntTpe)

but instead we were creating the following type:

MethodType(List(<symbol of Array>), IntTpe)

which is very wrong. As a consequence, GenICode would see our
method to have the following signature:

def m(xs: Array[T]): Int

for undefined T.

I checked other places in Delambdafy for the same mistake and it seems
like everywhere else it's ok.